TEMPORARY: Force full page reload instead of morph
This is a diagnostic change to verify the data is being updated correctly.
Instead of trying to morph the DOM, we'll force a full page reload.

If this works (UI updates), then we know:
- Database updates are working
- Instance variables are being set correctly
- Problem is ONLY with the DOM morphing mechanism

If this DOESN'T work, then something else is wrong.

This should NOT be the final solution - just helps us diagnose.